requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.7.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
redis==5.0.1
python-multipart==0.0.6
pymupdf==1.23.7
sqlalchemy==2.0.23
cx-Oracle==8.3.0
uvloop>=0.14.0
httptools>=0.5.0
